server/pyproject.toml
[tool.poetry]
name = "anyknowage-server"
version = "0.1.0"
description = ""
authors = ["phillychi3 <phillychi3@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-socketio = "^5.11.2"
loguru = "^0.7.2"
haystack-ai = "^2.1.2"
qdrant-haystack = "^3.5.0"
fastembed-haystack = "^1.1.0"
attrs = "^23.2.0"
pymongo = "^4.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F"]
ignore = ["E501","F401","E711","E712"]
fixable = ["A", "B", "C", "D", "E", "F" ]
unfixable = []